Isolato taxRule

+++ src/main/kotlin/cgm/hexagonal/salestaxesKata/domain/models/SaleArticle.kt
@@ -1,6 +1,6 @@
 package cgm.hexagonal.salestaxesKata.domain.models
 
-data class SaleArticle(val article: Article, val quantity: Double) {
+data class SaleArticle(val article: Article, val quantity: Int) {
     var taxedPrice: Double = 0.0
     var tax: Double = 0.0
 }
